+++ cognite/neat/app/api/context_manager/__init__.py
@@ -1,0 +1,3 @@
+from cognite.neat.app.api.context_manager.manager import lifespan
+
+__all__ = ["lifespan"]
