+++ cognite/neat/graph/extractors/rdf_to_graph.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from rdflib import Graph, Namespace
 
-from cognite.neat.app.api.configuration import DEFAULT_NAMESPACE, DEFAULT_URI, PREFIXES
+from cognite.neat.constants import DEFAULT_NAMESPACE, DEFAULT_URI, PREFIXES
 
 __all__ = ["rdf_file_to_graph"]
