+++ cognite/neat/rules/__init__.py
@@ -1,5 +1,5 @@
 from cognite.neat.rules import models
-from cognite.neat.rules._parser import (
+from cognite.neat.rules.parser import (
     parse_rules_from_excel_file,
     parse_rules_from_google_sheet,
     parse_rules_from_github_sheet,
