+++ tests/conftest.py
@@ -5,7 +5,7 @@ from cognite.neat import rules
 from cognite.neat.graph import loaders, extractors
 from cognite.neat.graph.extractors import NeatGraphStore
 from cognite.neat.graph.extractors.mocks import generate_triples
-from cognite.neat.rules._parser import RawTables, read_excel_file_to_table_by_name
+from cognite.neat.rules.parser import RawTables, read_excel_file_to_table_by_name
 from cognite.neat.rules.importer.ontology2excel import owl2excel
 from cognite.neat.rules.models import TransformationRules
 from cognite.neat.graph.transformations.transformer import domain2app_knowledge_graph
