+++ tests/graph/transformations/query_generator/test_sparql.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from cognite.neat.app.api.configuration import PREFIXES
+from cognite.neat.constants import PREFIXES
 from cognite.neat.graph.extractors import NeatGraphStore
 from cognite.neat.graph.transformations.query_generator.sparql import build_sparql_query
