+++ tests/rules/test_to_rdf_path.py
@@ -5,7 +5,7 @@ from pydantic import ValidationError
 import pprint
 from IPython.display import Markdown, display
 
-from cognite.neat.app.api.configuration import PREFIXES
+from cognite.neat.constants import PREFIXES
 from cognite.neat.graph.extractors import NeatGraphStore
 from cognite.neat.graph.transformations.query_generator import build_sparql_query
 from cognite.neat.rules.models import Property
